[VhlsLLVM] refactorized by splitting files
