+++ lib/llvm/Transforms/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_llvm_library(VhlsLLVMRewriter MODULE
   VhlsLLVMRewriter.cc
+  MemRefToArray.cc
+  Utils.cc
 
   PLUGIN_TOOL
   opt
